Feat(#16): 전/월세 데이터 필터링 기능 추가
